requirements.txt
asgiref==3.3.0
cffi==1.14.3
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
pycparser==2.20
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
